% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonApplyTagTypeWeights}
\alias{SMRMonApplyTagTypeWeights}
\title{Apply tag type weights}
\usage{
SMRMonApplyTagTypeWeights(smrObj, weights, default = 1)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{weights}{A list/vector of weights to be applied.}

\item{default}{The weight to be used for weights not specified in weights.}
}
\value{
An SMRMon object.
}
\description{
Applies the weights of tag types of a sparse matrix recommender object.
}
\details{
If \code{weights} does not have names it is going to replicated to match
the length of \code{smr$TagTypes}.
If \code{weights} has names the missing tag types are (if any)
are set to have the value \code{default}.
}
