% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonApplyTagWeights}
\alias{SMRMonApplyTagWeights}
\title{Apply tag weights}
\usage{
SMRMonApplyTagWeights(smrObj, weights, default = 1)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{weights}{A list/vector of weights to be applied.}

\item{default}{The weight to be used for weights not specified in weights.}
}
\value{
An SMRMon object.
}
\description{
Multiplies the weights of tags of a sparse matrix recommender object.
}
\details{
If \code{weights} does not have names it is going to replicated to match
the length of \code{ncol(smr$M)}.
If \code{weights} has names the missing tag types are (if any)
are set to have the value \code{default}.
}
