% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonApplyTermWeightFunctions}
\alias{SMRMonApplyTermWeightFunctions}
\title{Apply term-weight functions.}
\usage{
SMRMonApplyTermWeightFunctions(smrObj, globalWeightFunction = "IDF",
  localWeightFunction = "None", normalizerFunction = "Cosine")
}
\arguments{
\item{smrObj}{A SMRMon object.}

\item{globalWeightFunction}{Global weight function.}

\item{localWeightFunction}{Local weight function.}

\item{normalizerFunction}{Normalizer function.}
}
\value{
A SMRMon object.
}
\description{
Re-weight the entries of the recommender matrix (per tag type.)
}
\details{
This function calls
\code{\link{SparseMatrixRecommender::SMRApplyTermWeightFunctions}} for each
sub-matrix of \code{smrObj$M}.
}
