% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonBatchRecommend}
\alias{SMRMonBatchRecommend}
\title{Batch recommend}
\usage{
SMRMonBatchRecommend(
  smrObj,
  data = NULL,
  nrecs = 12,
  removeHistoryQ = FALSE,
  normalizeQ = FALSE,
  targetColumnName = "ProfileID"
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{data}{A matrix of profiles that has the same number of columns as
\code{smrObj$M}, or a vector of items (i.e. rownames of \code{smrObj$M}.)
If NULL then \code{smrObj$M} is used.}

\item{nrecs}{Number of recommendations to be returned.
If NULL all non-zero score recommendations are returned.}

\item{removeHistoryQ}{Should the history be removed from the recommendations?}

\item{normalizeQ}{Should the scores be normalized by the maximum score or not?}

\item{targetColumnName}{The column name for the \code{data} rownames in the result data frame.
(I.e. the column in the recommendations result data frame that has \code{rownames(mat)} as values.)}
}
\value{
A SMRMon object
}
\description{
For each row of a specified matrix of profiles compute recommendations.
}
\details{
The recommendations result is a
data frame with columns \code{c(targetColumnName "Score", "Index", smrObj$ItemColumnName)};
assigned to \code{smrObj$Value}. The following steps are taken:
(1) Make sure the data argument is \code{"dgCMatrix"}.
(2) Compute matrix-matrix product for the recommendations.
(3) If specified, remove history.
(4) Keep the specified top number of recommendations only.
(5) If specified, normalize per data element the scores by the maximum score.
(6) Return results.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{SMRMonGetTopRecommendations}()},
\code{\link{SMRMonJoinAcross}()},
\code{\link{SMRMonProveByHistory}()},
\code{\link{SMRMonProveByMetadata}()},
\code{\link{SMRMonProve}()},
\code{\link{SMRMonRecommendByProfile}()},
\code{\link{SMRMonRecommend}()}
}
\concept{Recommendations computation functions}
