% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonClassifyByProfile}
\alias{SMRMonClassifyByProfile}
\title{Classification with a profile}
\usage{
SMRMonClassifyByProfile(smrObj, tagType, profile, nTopNNs = NULL,
  voting = FALSE, dropZeroScoredLabels = TRUE,
  maxNumberOfLabels = NULL, normalizeQ = TRUE)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{tagType}{Tag type for which the classification is done.}

\item{profile}{(Scored) profile tags.
A data frame with columns \code{c("Score", "Tag")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{nTopNNs}{Number of top nearest neighbors to be used in the derive the classification.
If NULL one hundred top nearest neighbors are used.
(More precisely \code{min( 100, nrow(smrObj$M) )}.)}

\item{voting}{Should simple voting be used or a weighted sum?}

\item{maxNumberOfLabels}{The maximum number of labels to be returned;
if NULL all found labels are returned.}

\item{normalizeQ}{Should the scores be normalized?
(By dividing by the maximum score.)}
}
\value{
An SMRMon object.
}
\description{
Classify a profile data frame into the column names of a tag type sub-matrix.
}
\details{
The classification result is a list of scored tags that is assigned
to \code{smrObj$Value}.
This function is based in \code{\link{SMRClassifyByProfileVector}}.
}
