% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonCreate}
\alias{SMRMonCreate}
\title{Make an SMRMon object from a transactions data frame}
\usage{
SMRMonCreate(smrObj, data = SMRMonTakeData(smrObj),
  tagTypes = names(data)[-1], itemColumnName = names(data)[1], ...)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{data}{The transactions data frame.}

\item{tagTypes}{The name of the column containing the categorical tags.}

\item{itemColumnName}{The name of the column containing the unique items.}

\item{...}{Additional parameters for \code{\link{SMRMonCreateFromMatrices}}.}
}
\value{
SMR object.
}
\description{
Creates a sparse matrix recommender from transactions data and a list of tag types,
and makes is an SMRMon object.
}
\details{
An S3 object is returned that is list with class attribute set to "SMR".
}
\seealso{
Other Creation functions: \code{\link{SMRMonCreateFromMatrices}}
}
\concept{Creation functions}
