% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonEcho}
\alias{SMRMonEcho}
\title{Echo.}
\usage{
SMRMonEcho(smrObj, x)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{x}{Object to be echoed.}
}
\value{
An SMRMon object
}
\description{
Echoes the given argument.
}
\seealso{
Other Echo functions: 
\code{\link{SMRMonEchoFunctionValue}()},
\code{\link{SMRMonEchoValue}()}
}
\concept{Echo functions}
