% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonEchoFunctionValue}
\alias{SMRMonEchoFunctionValue}
\title{Function application to monad's value.}
\usage{
SMRMonEchoFunctionValue(smrObj, f)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{f}{A function to be applied to \code{smrObj$Value}.}
}
\value{
A SMRMon object.
}
\description{
Apply a function to the "Value" element/member of the monad object.
}
\details{
Prints \code{f(smrObj$Value)}.
}
