% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonEchoValue}
\alias{SMRMonEchoValue}
\title{Echo monad's value.}
\usage{
SMRMonEchoValue(smrObj)
}
\arguments{
\item{smrObj}{An SMRMon object.}
}
\value{
An SMRMon object
}
\description{
Echoes the monad object value.
}
\seealso{
Other Echo functions: 
\code{\link{SMRMonEchoFunctionValue}()}
}
\concept{Echo functions}
