% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonEnhanceWithNearestNeighbors}
\alias{SMRMonEnhanceWithNearestNeighbors}
\title{Enhance with nearest neighbors sub-matrix}
\usage{
SMRMonEnhanceWithNearestNeighbors(
  smrObj,
  tagTypes = NULL,
  numberOfNearestNeighbors = 20,
  batchSize = NULL
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{tagTypes}{Tag types to use make compute the similarities.}

\item{numberOfNearestNeighbors}{Number of nearest neighbors per item.}

\item{batchSize}{Batch size to use for the computations}
}
\value{
An SMRMon object.
}
\description{
Enhances the given SMR object with nearest neighbors sub-matrix.
}
