% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\docType{data}
\name{SMRMonFailureSymbol}
\alias{SMRMonFailureSymbol}
\title{Failure symbol for SMRMon.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
SMRMonFailureSymbol
}
\description{
Failure symbol for the monad SMRMon.
}
\keyword{datasets}
