% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonFilterByProfile}
\alias{SMRMonFilterByProfile}
\title{Filter by profile}
\usage{
SMRMonFilterByProfile(smrObj, profile, type = "intersection")
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{profile}{A profile specification used to filter with.}

\item{type}{The type of filtering one of "union" or "intersection".}
}
\value{
An SMRMon object.
}
\description{
Applies a profile filter to the rows of the recommendation matrix.
}
\details{
The result is a vector of scored items that is assigned to \code{smrObj$Value}.
If \code{type} is "union" each item that has at least one of the tags in
\code{profile} is in the result.
(Essentially, that is the same as taking all non-zero score recommendations by profile.)
If \code{type} is "intersection" each item in the result
has all tags in \code{profile}.
}
