% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonFindAnomalies}
\alias{SMRMonFindAnomalies}
\title{Find anomalies by nearest neighbors distances}
\usage{
SMRMonFindAnomalies(
  smrObj,
  data = NULL,
  numberOfNearestNeighbors = 12,
  aggregationFunction = mean,
  thresholdsIdentifier = NULL,
  normalizeQ = FALSE,
  property = "RowNames",
  useBatchRecommendationQ = TRUE
)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{data}{Data vectors to be evaluated as anomalous or not.
If NULL the \code{smrObj$M} is used.}

\item{numberOfNearestNeighbors}{Number of nearest neighbors.}

\item{aggregationFunction}{Aggregation function for the nearest neighbor distances.
If NULL \code{mean} is used.}

\item{thresholdsIdentifier}{Outlier thresholds identifier.
If NULL then
\code{function(x) OutlierIdentifiers::BottomOutliersOnlyThresholds( OutlierIdentifiers::HampelIdentifierParameters(x) )}
is used.}

\item{normalizeQ}{Should each recommendation scores be normalized or not?}

\item{property}{A string.
One of \code{c("Similarities", "SparseMatrix", "RowNames", "OutlierThresholds", "Properties")}}

\item{useBatchRecommendationQ}{Should batch recommendation be used or not?}
}
\value{
An SMRMon object.
}
\description{
Find anomalies by nearest neighbors distances distributions.
}
\details{
The computation result is assigned to \code{smrObj$Value}.
The computation steps follow.
(1) For each row of the given data find the specified number of Nearest Neighbors (NNs) in the recommender matrix.
(2) For each row aggregate the NNs scores with the specified aggregation function.
(3) Find outlier thresholds for the list of aggregated values.
(4) Identify the outliers by using the outlier thresholds.
}
