% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonGetHistoryDataFrame}
\alias{SMRMonGetHistoryDataFrame}
\title{From a history specification into a data frame}
\usage{
SMRMonGetHistoryDataFrame(
  smrObj,
  history,
  functionName = "SMRMonGetHistoryDataFrame",
  warningQ = TRUE
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{history}{History specification.
A data frame with columns \code{c("Rating", "Item")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{functionName}{A string that is a name of this function or a delegating function.}

\item{warningQ}{Should a warning be issued if \code{history} is of unknown type?}
}
\value{
A SMRMon object
}
\description{
Transforms a history specification into a data frame.
}
\details{
The result data frame is with columns "Score", \code{smr$ItemColumnName};
assigned to \code{smrObj$Value}.
If \code{history = NULL} then \code{smrObj$Value} is used.
}
