% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonGetMatrixProperty}
\alias{SMRMonGetMatrixProperty}
\title{Get a member of a SMRMon object}
\usage{
SMRMonGetMatrixProperty(smrObj, property, tagType = NULL)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{property}{A string.}

\item{tagType}{Tag type string (sub-matrix name) to get the property of.
If NULL the whole recommendation matrix, \code{smrObj$M}, is used.}
}
\value{
SMR object.
}
\description{
Places an SMRMon object member to be the pipeline value.
}
\details{
This function simplifies certain workflows.
}
\seealso{
Other Get functions: 
\code{\link{SMRMonGetProperty}()}
}
\concept{Get functions}
