% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonGetProfileDataFrame}
\alias{SMRMonGetProfileDataFrame}
\title{From a profile specification into a data frame}
\usage{
SMRMonGetProfileDataFrame(smrObj, profile,
  functionName = "SMRMonGetProfileDataFrame")
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{profile}{Profile specification.
A data frame with columns \code{c("Score", "Tag")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A SMRMon object
}
\description{
Transforms a profile specification into a data frame.
}
\details{
The result data frame is with columns "Score", \code{smr$ItemColumnName};
assigned to \code{smrObj$Value}.
If \code{profile = NULL} then \code{smrObj$Value} is used.
}
