% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonGetProperty}
\alias{SMRMonGetProperty}
\title{Get a member of a SMRMon object}
\usage{
SMRMonGetProperty(smrObj, property)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{property}{A string.}
}
\value{
SMR object.
}
\description{
Places an SMRMon object member to be the pipeline value.
}
\details{
This function simplifies certain workflows.
}
\seealso{
Other Get functions: 
\code{\link{SMRMonGetMatrixProperty}()}
}
\concept{Get functions}
