% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonGetTopRecommendations}
\alias{SMRMonGetTopRecommendations}
\title{Get top recommendations}
\usage{
SMRMonGetTopRecommendations(smrObj, spec = NULL, nrecs = 12, ...)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{spec}{A history or profile specification.
If NULL \code{smrObj$Value} is (tried to be) used.}
}
\value{
SMR object.
}
\description{
Computes the top recommendations for a history or profile argument
or monad's value.
}
\details{
This function simplifies certain workflows.
(Very inefficient implementation at this point.)
}
\seealso{
Other Recommendations computation functions: 
\code{\link{SMRMonJoinAcross}()},
\code{\link{SMRMonProveByHistory}()},
\code{\link{SMRMonProveByMetadata}()},
\code{\link{SMRMonProve}()},
\code{\link{SMRMonRecommendByProfile}()},
\code{\link{SMRMonRecommend}()}
}
\concept{Recommendations computation functions}
