% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonJoinAccross}
\alias{SMRMonJoinAccross}
\title{Extend recommendations with extra data}
\usage{
SMRMonJoinAccross(smrObj, data, by = smrObj$ItemColumnName)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{data}{A data frame.}
}
\value{
An SMRMon object.
}
\description{
Joins a recommendations data with a given data frame.
}
\details{
The argument \code{data} is expected to have a column with name
\code{smrObj$ItemColumnName} or a inner join column that is specified
with the argument \code{by}.
The name of the function corresponds a Mathematica function (\code{JoinAccross}).
}
\seealso{
Other Recommendations computation functions: \code{\link{SMRMonRecommendByProfile}},
  \code{\link{SMRMonRecommend}}
}
\concept{Recommendations computation functions}
