% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonMakeTagTypeRecommender}
\alias{SMRMonMakeTagTypeRecommender}
\title{Make tag type recommender}
\usage{
SMRMonMakeTagTypeRecommender(
  smrObj,
  tagTypeTo,
  nTopTags = 1,
  tagTypes = NULL,
  tagSelectionCriteria = NULL,
  ...
)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{tagTypeTo}{Tag type to make a recommender for.}

\item{nTopTags}{Number of top tags from \code{tagTypeTo} when making item-tag
replacements.}

\item{tagTypes}{A vector tag types (strings) to make the data frame with.
If NULL all tag types are used. Passed to \code{\link{SMRMatricesToLongForm}}.}

\item{tagSelectionCriteria}{Tag selection criteria.
If a positive integer for each tag type the number of \code{tagSelectionCriteria} top tags are taken.
If a function that function is expected to give a list of booleans for a given list of tag weights.
The tags with weights that correspond to TRUE are selected.}

\item{...}{Additional arguments for \code{\link{SMRMatricesToLongForm}} or \code{\link{SMRCreateFromMatrices}}.}
}
\value{
A sparse matrix recommender
}
\description{
Converts the recommender into a recommender for one of the tag types.
}
\details{
This monad function calls the function \code{\link{SMRToMetadataRecommender}}.
The following steps are taken.
(1) The long form of the recommender is made.
(2) The items are replaced with the top tags of \code{tagTypeTo}.
(3) A new recommender is created with items that are the tags of \code{tagTypeTo}.
}
