% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonProfile}
\alias{SMRMonProfile}
\title{Profile for a history specification}
\usage{
SMRMonProfile(smrObj, history)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{history}{History specification.
A data frame with columns \code{c("Rating", "Item")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}
}
\value{
A SMRMon object
}
\description{
Find the profile that corresponds to a history specification.
}
\details{
The recommendations result is a
data frame with columns "Score", "Index", "Tag";
assigned to \code{smrObj$Value}.
}
