% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonProfileSpecificationQ}
\alias{SMRMonProfileSpecificationQ}
\title{Is the argument a profile specification.}
\usage{
SMRMonProfileSpecificationQ(smrObj, spec,
  functionName = "SMRMonProfileSpecificationQ", logicalResultQ = FALSE)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{spec}{Data to be tested. If NULL \code{smrObj$Value} is tested.}

\item{functionName}{A string that is a name of this function or a delegating function.}

\item{logicalResultQ}{Should the result be a logical value?}
}
\value{
SMR object.
}
\description{
Tests is the argument a profile.
If \code{smrObj$M} is a (recommendation) sparse matrix then the profile tags
are tested are they known columns of \code{smrObj$M}.
}
\details{
This function simplifies certain workflows.
}
\concept{Predicate functions}
