% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonProve}
\alias{SMRMonProve}
\title{Prove a recommendation to a profile or history}
\usage{
SMRMonProve(
  smrObj,
  spec,
  items,
  type = "metadata",
  normalizeScoresQ = TRUE,
  style = "intersection",
  outlierIdentifierParameters = NULL,
  warningQ = TRUE
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{spec}{A specification that is a profile or a history.}

\item{items}{One or more item-indexes or item-names to make proofs for.}

\item{type}{One of "metadata" or "history".
Must be a row index or row name of \code{smrObj$M}.}

\item{normalizeScoresQ}{Should the proof scores be normalized or not?}

\item{style}{Proof style derivation; one of "intersection", "multiplication".}

\item{outlierIdentifierParameters}{Outlier identifier parameters or a parameters finding function.
If NULL all tags are returned.}

\item{warningQ}{Should warnings be given?}
}
\value{
An SMRMon object.
}
\description{
If \code{proofType = "metadata"} then for a given profile and a recommended item
finds scored metadata tags that appear in both the given profile and the profile
of the recommended item.
If \code{proofType = "history"} then for a given history and a recommended item
finds the items of the history that are the closest to recommended item.
}
\details{
The result is assigned to \code{smrObj$Value}.
For some data and outlier identifier parameters the proofs data frame might turn out empty.
In those cases warning a given if \code{warningQ = TRUE}.
The argument \code{style} has an effect only when \code{type = "metadata"}.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{SMRMonBatchRecommend}()},
\code{\link{SMRMonGetTopRecommendations}()},
\code{\link{SMRMonJoinAcross}()},
\code{\link{SMRMonProveByHistory}()},
\code{\link{SMRMonProveByMetadata}()},
\code{\link{SMRMonRecommendByProfile}()},
\code{\link{SMRMonRecommend}()}
}
\concept{Recommendations computation functions}
