% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonProveByHistory}
\alias{SMRMonProveByHistory}
\title{Prove a recommendation to a history.}
\usage{
SMRMonProveByHistory(
  smrObj,
  history,
  item,
  normalizeScoresQ = TRUE,
  outlierIdentifierParameters = NULL,
  warningQ = TRUE
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{history}{A history specification.}

\item{item}{An item index or name to make proofs for.
Must be a row index or row name of \code{smrObj$M}.}

\item{normalizeScoresQ}{Should the proof scores be normalized or not?}

\item{outlierIdentifierParameters}{Outlier identifier parameters or parameters finding function.
If NULL all items are returned.}

\item{warningQ}{Should warnings be given?}
}
\value{
An SMRMon object.
}
\description{
Finds the items of the history that are the closest to a recommendation.
}
\details{
The result is a data frame with columns names
"Score", "Index", and \code{smrObj$ItemColumnName}
and it is assigned to \code{smrObj$Value}.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{SMRMonGetTopRecommendations}()},
\code{\link{SMRMonJoinAcross}()},
\code{\link{SMRMonProveByMetadata}()},
\code{\link{SMRMonProve}()},
\code{\link{SMRMonRecommendByProfile}()},
\code{\link{SMRMonRecommend}()}
}
\concept{Recommendations computation functions}
