% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonRecommend}
\alias{SMRMonRecommend}
\title{Compute recommendations}
\usage{
SMRMonRecommend(
  smrObj,
  history,
  nrecs = 12,
  removeHistoryQ = FALSE,
  normalizeQ = FALSE,
  warningQ = TRUE
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{history}{History specification.
A data frame with columns \code{c("Rating", "Item")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{nrecs}{Number of recommendations to be returned.
If NULL all non-zero score items are returned.}

\item{removeHistoryQ}{Should the history be removed from the recommendations?}

\item{normalizeQ}{Should the scores be normalized by the maximum score or not?}

\item{warningQ}{Should a warning be issued if \code{history} is of unknown type?}
}
\value{
A SMRMon object
}
\description{
Recommend items based on history of consumption.
}
\details{
The recommendations result is a
data frame with columns "Score", "Index", \code{smrObj$ItemColumnName};
assigned to \code{smrObj$Value}.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{SMRMonBatchRecommend}()},
\code{\link{SMRMonGetTopRecommendations}()},
\code{\link{SMRMonJoinAcross}()},
\code{\link{SMRMonProveByHistory}()},
\code{\link{SMRMonProveByMetadata}()},
\code{\link{SMRMonProve}()},
\code{\link{SMRMonRecommendByProfile}()}
}
\concept{Recommendations computation functions}
