% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonRetrievalByProfileStatistics}
\alias{SMRMonRetrievalByProfileStatistics}
\title{Retrieval by profile statistics}
\usage{
SMRMonRetrievalByProfileStatistics(
  smrObj,
  focusTag,
  focusTagType,
  profileTagTypes,
  nrecs = 100
)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{focusTag}{The tag to be tested.}

\item{focusTagType}{The tag type of the tag to be tested.}

\item{profileTagTypes}{A character vector of tag types.}

\item{nrecs}{Number of recommendations.}
}
\value{
An SMRMon object.
}
\description{
Computes the Top-K statistics between the items that have a given focus tag
and the profile recommendations for that focus tag (as a profile.)
}
\details{
The computation result is assigned to \code{smrObj$Value}.
The computation steps follow.
(1) Find all items that have \code{focustTag}.
(2) Find the profile of the focus tag items.
(3) Find recommendations with the profile according to \code{max(nrecs)}.
(4) Find how many focus tag items (of step 1) are present in the recommendations (of step 3).
}
