% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonRetrieveByQueryElements}
\alias{SMRMonRetrieveByQueryElements}
\title{Retrieve by query elements}
\usage{
SMRMonRetrieveByQueryElements(
  smrObj,
  should = NULL,
  must = NULL,
  mustNot = NULL,
  mustType = "intersection",
  mustNotType = "union"
)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{should}{A profile specification used to recommend with.}

\item{must}{A profile specification used to filter with.
The items in the result must have the tags in \code{must}.}

\item{mustNot}{A profile specification used to filter with.
The items in the result must not have the tags in \code{mustNot}.}

\item{mustType}{The type of filtering with the must tags; one of "union" or "intersection".}

\item{mustNotType}{The type of filtering with the must not tags; one of "union" or "intersection".}
}
\value{
An SMRMon object.
}
\description{
Applies a profile filter to the rows of the recommendation matrix.
}
\details{
The result is assigned to \code{smrObj$Value}.
This function is based on \code{\link{SMRMonRecommendByProfile}} ("should")
and \code{\link{SMRMonFilterByProfile}} ("must" and "must not").
}
