% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonSetData}
\alias{SMRMonSetData}
\title{Set Data.}
\usage{
SMRMonSetData(smrObj, Data)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{Data}{An object member to be set.}
}
\value{
An SMRMon object.
}
\description{
Sets Data into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{SMRMonSetItemColumnName}},
  \code{\link{SMRMonSetItemToIndexRules}},
  \code{\link{SMRMonSetM01}}, \code{\link{SMRMonSetM}},
  \code{\link{SMRMonSetTagToIndexRules}},
  \code{\link{SMRMonSetTagTypeRange}},
  \code{\link{SMRMonSetTagTypes}},
  \code{\link{SMRMonSetValue}},
  \code{\link{SMRMonTakeData}},
  \code{\link{SMRMonTakeItemColumnName}},
  \code{\link{SMRMonTakeItemToIndexRules}},
  \code{\link{SMRMonTakeM01}}, \code{\link{SMRMonTakeM}},
  \code{\link{SMRMonTakeTagToIndexRules}},
  \code{\link{SMRMonTakeTagTypeRange}},
  \code{\link{SMRMonTakeTagTypes}},
  \code{\link{SMRMonTakeValue}}
}
\concept{Set/Take functions}
