% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonSetItemColumnName}
\alias{SMRMonSetItemColumnName}
\title{Set ItemColumnName.}
\usage{
SMRMonSetItemColumnName(smrObj, ItemColumnName)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{ItemColumnName}{An object member to be set.}
}
\value{
An SMRMon object.
}
\description{
Sets ItemColumnName into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{SMRMonSetData}},
  \code{\link{SMRMonSetItemToIndexRules}},
  \code{\link{SMRMonSetM01}}, \code{\link{SMRMonSetM}},
  \code{\link{SMRMonSetTagToIndexRules}},
  \code{\link{SMRMonSetTagTypeRanges}},
  \code{\link{SMRMonSetTagTypes}},
  \code{\link{SMRMonSetValue}},
  \code{\link{SMRMonTakeData}},
  \code{\link{SMRMonTakeItemColumnName}},
  \code{\link{SMRMonTakeItemToIndexRules}},
  \code{\link{SMRMonTakeM01}}, \code{\link{SMRMonTakeM}},
  \code{\link{SMRMonTakeTagToIndexRules}},
  \code{\link{SMRMonTakeTagTypeRanges}},
  \code{\link{SMRMonTakeTagTypes}},
  \code{\link{SMRMonTakeValue}}
}
\concept{Set/Take functions}
