% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonSetTagToIndexRules}
\alias{SMRMonSetTagToIndexRules}
\title{Set TagToIndexRules.}
\usage{
SMRMonSetTagToIndexRules(smrObj, TagToIndexRules)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{TagToIndexRules}{An object member to be set.}
}
\value{
An SMRMon object.
}
\description{
Sets TagToIndexRules into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{SMRMonSetData}()},
\code{\link{SMRMonSetItemColumnName}()},
\code{\link{SMRMonSetItemToIndexRules}()},
\code{\link{SMRMonSetM01}()},
\code{\link{SMRMonSetM}()},
\code{\link{SMRMonSetTagTypeRanges}()},
\code{\link{SMRMonSetTagTypes}()},
\code{\link{SMRMonSetValue}()},
\code{\link{SMRMonTakeData}()},
\code{\link{SMRMonTakeItemColumnName}()},
\code{\link{SMRMonTakeItemToIndexRules}()},
\code{\link{SMRMonTakeM01}()},
\code{\link{SMRMonTakeM}()},
\code{\link{SMRMonTakeTagToIndexRules}()},
\code{\link{SMRMonTakeTagTypeRanges}()},
\code{\link{SMRMonTakeTagTypes}()},
\code{\link{SMRMonTakeValue}()}
}
\concept{Set/Take functions}
