% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonSetTagTypeRanges}
\alias{SMRMonSetTagTypeRanges}
\title{Set TagTypeRanges.}
\usage{
SMRMonSetTagTypeRanges(smrObj, TagTypeRanges)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{TagTypeRanges}{An object member to be set.}
}
\value{
An SMRMon object.
}
\description{
Sets TagTypeRanges into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{SMRMonSetData}},
  \code{\link{SMRMonSetItemColumnName}},
  \code{\link{SMRMonSetItemToIndexRules}},
  \code{\link{SMRMonSetM01}}, \code{\link{SMRMonSetM}},
  \code{\link{SMRMonSetTagToIndexRules}},
  \code{\link{SMRMonSetTagTypes}},
  \code{\link{SMRMonSetValue}},
  \code{\link{SMRMonTakeData}},
  \code{\link{SMRMonTakeItemColumnName}},
  \code{\link{SMRMonTakeItemToIndexRules}},
  \code{\link{SMRMonTakeM01}}, \code{\link{SMRMonTakeM}},
  \code{\link{SMRMonTakeTagToIndexRules}},
  \code{\link{SMRMonTakeTagTypeRanges}},
  \code{\link{SMRMonTakeTagTypes}},
  \code{\link{SMRMonTakeValue}}
}
\concept{Set/Take functions}
