% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonSetTagTypes}
\alias{SMRMonSetTagTypes}
\title{Set TagTypes.}
\usage{
SMRMonSetTagTypes(smrObj, TagTypes)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{TagTypes}{An object member to be set.}
}
\value{
An SMRMon object.
}
\description{
Sets TagTypes into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{SMRMonSetData}},
  \code{\link{SMRMonSetItemColumnName}},
  \code{\link{SMRMonSetItemToIndexRules}},
  \code{\link{SMRMonSetM01}}, \code{\link{SMRMonSetM}},
  \code{\link{SMRMonSetTagToIndexRules}},
  \code{\link{SMRMonSetTagTypeRange}},
  \code{\link{SMRMonSetValue}},
  \code{\link{SMRMonTakeData}},
  \code{\link{SMRMonTakeItemColumnName}},
  \code{\link{SMRMonTakeItemToIndexRules}},
  \code{\link{SMRMonTakeM01}}, \code{\link{SMRMonTakeM}},
  \code{\link{SMRMonTakeTagToIndexRules}},
  \code{\link{SMRMonTakeTagTypeRange}},
  \code{\link{SMRMonTakeTagTypes}},
  \code{\link{SMRMonTakeValue}}
}
\concept{Set/Take functions}
