% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonSetValue}
\alias{SMRMonSetValue}
\title{Set the value in a SMRMon object.}
\usage{
SMRMonSetValue(smrObj, value)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{value}{The new value.}
}
\value{
A SMRMon object.
}
\description{
Sets the value in a SMRMon monad object.
}
\details{
Assigns \code{value} to \code{smrObj$Value}.
}
\seealso{
Other Set/Take functions: 
\code{\link{SMRMonSetData}()},
\code{\link{SMRMonSetItemColumnName}()},
\code{\link{SMRMonSetItemToIndexRules}()},
\code{\link{SMRMonSetM01}()},
\code{\link{SMRMonSetM}()},
\code{\link{SMRMonSetTagToIndexRules}()},
\code{\link{SMRMonSetTagTypeRanges}()},
\code{\link{SMRMonSetTagTypes}()},
\code{\link{SMRMonTakeData}()},
\code{\link{SMRMonTakeItemColumnName}()},
\code{\link{SMRMonTakeItemToIndexRules}()},
\code{\link{SMRMonTakeM01}()},
\code{\link{SMRMonTakeM}()},
\code{\link{SMRMonTakeTagToIndexRules}()},
\code{\link{SMRMonTakeTagTypeRanges}()},
\code{\link{SMRMonTakeTagTypes}()},
\code{\link{SMRMonTakeValue}()}
}
\concept{Set/Take functions}
