% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonSummarizeItem}
\alias{SMRMonSummarizeItem}
\title{Summarize an item}
\usage{
SMRMonSummarizeItem(smrObj, item)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{item}{A string.
(That is one of \code{rownames(smrObj$M)}.)}
}
\value{
A SMRMon object
}
\description{
Summarizes the recommender matrix data
for a given item.
(A row of the recommender matrix \code{smrObj$M}.)
}
\details{
The result is a list with named elements assigned to \code{smrObj$Value}.
}
