% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonTagNearestNeighbors}
\alias{SMRMonTagNearestNeighbors}
\title{Tag nearest neighbors}
\usage{
SMRMonTagNearestNeighbors(
  smrObj,
  tags,
  tagType,
  nrecs = 12,
  nrecsProfile = 100,
  normalizeQ,
  ...
)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{tags}{Tags for which nearest neighbors are found.}

\item{tagType}{The tag type of the nearest neighbors.}

\item{nrecs}{Number of nearest neighbors.}

\item{nrecsProfile}{Number of recommendations for finding the \code{tags} profile.}

\item{normalizeQ}{Should the scores be normalized?
(By dividing by the maximum score.)}

\item{...}{Additional arguments passed to \code{\link{SMRClassifyByProfileVector}}.}
}
\value{
An SMRMon object.
}
\description{
Find nearest neighbors for a given vector of tags.
}
\details{
The result is a list of scored tags that is assigned
to \code{smrObj$Value}.
This function is based in \code{\link{SMRClassifyByProfileVector}}.
The tags correspond to columns of the SMR object sparse matrix.
(The columns of that matrix are assumed to be unique.)
}
