% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonTakeM01}
\alias{SMRMonTakeM01}
\title{Take M01.}
\usage{
SMRMonTakeM01(smrObj, functionName = "SMRMonTakeM01")
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{SMRMonFailureSymbol}.
}
\description{
Takes M01 from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{SMRMonSetData}()},
\code{\link{SMRMonSetItemColumnName}()},
\code{\link{SMRMonSetItemToIndexRules}()},
\code{\link{SMRMonSetM01}()},
\code{\link{SMRMonSetM}()},
\code{\link{SMRMonSetTagToIndexRules}()},
\code{\link{SMRMonSetTagTypeRanges}()},
\code{\link{SMRMonSetTagTypes}()},
\code{\link{SMRMonSetValue}()},
\code{\link{SMRMonTakeData}()},
\code{\link{SMRMonTakeItemColumnName}()},
\code{\link{SMRMonTakeItemToIndexRules}()},
\code{\link{SMRMonTakeM}()},
\code{\link{SMRMonTakeTagToIndexRules}()},
\code{\link{SMRMonTakeTagTypeRanges}()},
\code{\link{SMRMonTakeTagTypes}()},
\code{\link{SMRMonTakeValue}()}
}
\concept{Set/Take functions}
