% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonUnit}
\alias{SMRMonUnit}
\title{Make a SMRMon Unit}
\usage{
SMRMonUnit(data = NULL)
}
\value{
An S3 class "SMR". In other words, a list with the attribute "class" set to "SMR".
}
\description{
Creates a monad object.
}
