% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonComputeTopK}
\alias{SMRMonComputeTopK}
\title{Top-K statistic computation}
\usage{
SMRMonComputeTopK(smrObj, testData, ks, ...)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{ks}{An integer vector with k's for the Top-K statistic.}

\item{...}{Additional arguments passed to \code{\link{SMRMonRecommend}}.}
}
\value{
An SMRMon object.
}
\description{
Computes the Top-K statistic for a data frame with the scored similarity pairs.
}
\details{
The computation result is assigned to \code{smrObj$Value}.
}
