% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonCreateFromMatrices}
\alias{SMRMonCreateFromMatrices}
\title{Creation an SMRMon object with a list of matrices.}
\usage{
SMRMonCreateFromMatrices(smrObj, matrices, tagTypes = names(matrices),
  itemColumnName = "Item", imposeSameRowNamesQ = TRUE,
  addTagTypesToColumnNamesQ = FALSE, sep = ":")
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{matrices}{A list of matrices to be spliced into a metadata matrix.}

\item{tagTypes}{Vector of matrix names.}

\item{itemColumnName}{The column name of recommender items (in data and recommendations).}

\item{imposeSameRowNamesQ}{Should the union of the row names be imposed on each matrix?}

\item{addTagTypesToColumnNamesQ}{Should the tag types be added as prefixes
to the column names of the corresponding sub-matrices?}

\item{sep}{Separator for the prefixes of the columns names.}
}
\value{
An SMRMon object.
}
\description{
Creates a sparse matrix recommender from a list of matrices and a corresponding list of tag types,
and makes is an SMRMon object.
}
\details{
An S3 object is returned that is list with class attribute set to "SMR".
}
\seealso{
Other Creation functions: \code{\link{SMRMonCreate}}
}
\concept{Creation functions}
