% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonFilterMatrix}
\alias{SMRMonFilterMatrix}
\title{Filter recommendation matrix rows}
\usage{
SMRMonFilterMatrix(smrObj, profile)
}
\arguments{
\item{smrObj}{A sparse matrix recommender.}

\item{profile}{A profile specification used to filter with.}
}
\value{
An SMRMon object.
}
\description{
Applies a profile filter to the rows of the recommendation matrix.
}
\details{
The matrix can be recovered with tag type
weights application, see \code{\link{SMRMonApplyTagTypeWeights}}.
}
