% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonGetLongFormData}
\alias{SMRMonGetLongFormData}
\title{Long form data frame representation.}
\usage{
SMRMonGetLongFormData(smrObj, tagTypesQ = TRUE)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{tagTypesQ}{Should the tag types be included or not?}
}
\value{
An SMRMon object.
}
\description{
Creates a long form data frame for the recommendation matrix of an SMRMon object.
}
\details{
The result data frame is assigned to \code{smrObj$Value}.
}
\concept{Data functions}
