% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonGetLongFormData}
\alias{SMRMonGetLongFormData}
\title{Long form data frame representation.}
\usage{
SMRMonGetLongFormData(
  smrObj,
  items = NULL,
  tagTypesQ = TRUE,
  summarizeTagsQ = FALSE,
  nTopTags = 5
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{items}{A character vector of items.
If NULL all elements of \code{rownames(smrObj$M)} are used.}

\item{tagTypesQ}{Should the tag types be included or not?}

\item{summarizeTagsQ}{Should the tags for each unique
item-and-tag-type pair be summarized or not?}

\item{nTopTags}{Number of top tags.
Used when \code{summarizeTagsQ = TRUE}.}
}
\value{
An SMRMon object.
}
\description{
Creates a long form data frame for the recommendation matrix of an SMRMon object.
}
\details{
The result data frame is assigned to \code{smrObj$Value}.
}
\concept{Data functions}
