% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonGetProfileVector}
\alias{SMRMonGetProfileVector}
\title{From a profile specification into a vector}
\usage{
SMRMonGetProfileVector(
  smrObj,
  profile,
  functionName = "SMRMonGetProfileVector",
  warningQ = TRUE
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{profile}{Profile specification.
A data frame with columns \code{c("Score", "Tag")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{functionName}{A string that is a name of this function or a delegating function.}

\item{warningQ}{Should a warning be issued if \code{profile} is of unknown type?}
}
\value{
A SMRMon object
}
\description{
Transforms a profile specification into a sparse matrix
with one row.
}
\details{
The result profile vector (sparse matrix) is
assigned to \code{smrObj$Value}.
If \code{profile = NULL} then \code{smrObj$Value} is used.
}
