% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonMemberPresenceCheck}
\alias{SMRMonMemberPresenceCheck}
\title{General member presence check.}
\usage{
SMRMonMemberPresenceCheck(
  smrObj,
  memberName,
  memberPrettyName = memberName,
  functionName = "",
  logicalResultQ = FALSE,
  warningQ = TRUE
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{memberName}{The name of the member to be checked.}

\item{memberPrettyName}{A pretty member name (for messages).}

\item{functionName}{The name of the delegating function.}

\item{logicalResultQ}{Should the result be a logical value?}

\item{warningQ}{Should a warning be issued or not?}
}
\value{
A logical value or an SMRMon object.
}
\description{
A general function for checking the presence of a data member in an SMRMon object.
}
