% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonOption}
\alias{SMRMonOption}
\title{Optional function application to monad's object.}
\usage{
SMRMonOption(smrObj, f)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{f}{A function to be applied to the monad object.}
}
\value{
A SMRMon object.
}
\description{
Optional function application to monad's object.
}
\details{
In general \code{f} should return a monad object,
but that is not enforced.
}
