% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonProfile}
\alias{SMRMonProfile}
\title{Profile for a history specification}
\usage{
SMRMonProfile(smrObj, history, tagTypesQ = FALSE, normalizeQ = FALSE)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{history}{History specification.
A data frame with columns \code{c("Rating", "Item")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{tagTypesQ}{Should the tag types be included or not?}

\item{normalizeQ}{Should the tag scores be normalized or not?}
}
\value{
A SMRMon object
}
\description{
Find the profile that corresponds to a history specification.
}
\details{
The profile data frame is assigned to \code{smrObj$Value}.
If \code{tagTypesQ = FALSE} then its are columns "Score", "Index", "Tag".
If \code{tagTypesQ = TRUE} then its are columns "Score", "Index", "TagType", "Tag".
}
