% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonProveByMetadata}
\alias{SMRMonProveByMetadata}
\title{Prove a recommendation to a profile using metadata.}
\usage{
SMRMonProveByMetadata(
  smrObj,
  profile,
  item,
  normalizeScoresQ = TRUE,
  style = "intersection"
)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{profile}{A profile specification.}

\item{item}{An item index or name to make proofs for.
Must be a row index or row name of \code{smrObj$M}.}

\item{normalizeScoresQ}{Should the proof scores be normalized or not?}

\item{style}{Proof style derivation; one of "intersection", "multiplication".}
}
\value{
An SMRMon object.
}
\description{
For a given profile and item finds scored metadata tags
that are appear in both the given profile and the profile of the given item.
}
\details{
The result is a data frame with columns names "Score", "Index", "Tag"
and it is assigned to \code{smrObj$Value}.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{SMRMonGetTopRecommendations}()},
\code{\link{SMRMonJoinAcross}()},
\code{\link{SMRMonProveByHistory}()},
\code{\link{SMRMonRecommendByProfile}()},
\code{\link{SMRMonRecommend}()}
}
\concept{Recommendations computation functions}
