% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonRecommend}
\alias{SMRMonRecommend}
\title{Compute recommendations}
\usage{
SMRMonRecommend(smrObj, history, nrecs = 12, removeHistoryQ = FALSE)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{history}{(Rated) history items.
A data frame with columns \code{c("Rating", "Item")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1.}

\item{nrecs}{Number of recommendations to be returned.}

\item{removeHistoryQ}{Should the history be removed from the recommendations?}
}
\value{
The recommendations result is a
data frame with columns "Score", "Index", \code{smr$ItemColumnName};
assigned to \code{smrObj$Value}.
}
\description{
Recommend items based on history of consumption.
}
\seealso{
Other Recommendations computation functions: \code{\link{SMRMonRecommendByProfile}}
}
\concept{Recommendations computation functions}
