% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMon.R
\name{SMRMonRecommendByProfile}
\alias{SMRMonRecommendByProfile}
\title{Recommendations using a profile}
\usage{
SMRMonRecommendByProfile(smrObj, profile, nrecs = 12, warningQ = TRUE)
}
\arguments{
\item{smrObj}{An SMRMon object.}

\item{profile}{Profile specification.
A data frame with columns \code{c("Score", "Tag")};
a numeric vector named elements, the names being items;
a character vector, the correspond ratings assumed all to be 1;
a sparse matrix with 1 column and a number of rows that equals \code{ncol(smrObj$M)}.}

\item{nrecs}{Number of recommendations to be returned.}

\item{warningQ}{Should a warning be issued if \code{profile} is of unknown type?}
}
\value{
A SMRMon object
}
\description{
Recommend items based on a sparse matrix and a specified profile indices and scores.
}
\details{
The recommendations result is a
data frame with columns "Score", "Index", \code{smr$ItemColumnName};
assigned to \code{smrObj$Value}.
}
\seealso{
Other Recommendations computation functions: 
\code{\link{SMRMonGetTopRecommendations}()},
\code{\link{SMRMonJoinAcross}()},
\code{\link{SMRMonProveByHistory}()},
\code{\link{SMRMonProveByMetadata}()},
\code{\link{SMRMonRecommend}()}
}
\concept{Recommendations computation functions}
